Facturacion funcional

+++ ProyectoIntegrador_2020/ExploradorClientes.vb
@@ -39,7 +39,6 @@ Public Class ExploradorClientes
     Private Sub btnNuevo_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles btnNuevo.Click
         Nuevo.lblTitulo.Text = "Nuevo Cliente"
         Nuevo.ShowDialog()
-        Me.Close()
     End Sub
 
 
@@ -219,8 +218,19 @@ Public Class ExploradorClientes
     End Sub
 
     Private Sub txtDireccion_KeyPress(ByVal sender As System.Object, ByVal e As System.Windows.Forms.KeyPressEventArgs) Handles txtDireccion.KeyPress
-        If Not Char.IsLetter(e.KeyChar) And Not (IsNumeric(e.KeyChar)) And Asc(e.KeyChar) <> 8 And Not Char.IsWhiteSpace(e.KeyChar) And Asc(e.KeyChar) <> 46 Then
-            e.Handled = True
+        If Not (IsNumeric(e.KeyChar)) And Asc(e.KeyChar) <> 8 And Asc(e.KeyChar) <> 46 And Asc(e.KeyChar) <> 44 Then
+            e.Handled = False
+
+            If Char.IsLetter(e.KeyChar) Then
+                e.Handled = False
+            ElseIf Char.IsControl(e.KeyChar) Then
+                e.Handled = False
+            ElseIf Char.IsSeparator(e.KeyChar) Then
+                e.Handled = False
+            Else
+                e.Handled = True
+            End If
+
         End If
     End Sub
 
@@ -244,9 +254,10 @@ Public Class ExploradorClientes
     Private Sub btnAgregarTel_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles btnAgregarTel.Click
         ConfirmacionMensaje.soloNumBool = True
         resultadosTxt = ConfirmacionMensaje.entradaDatos("Nuevo número telfónico:")
-
-        consultas.consultaHide("INSERT INTO telefonoCliente (numeroTel, idCliente) VALUES ('" & resultadosTxt & "'," & idCliente & ");")
-        ActualizarTablaTelefono()
+        If Not ConfirmacionMensaje.resultadoTxt.Equals("") And ConfirmacionMensaje.resultado = 1 Then
+            consultas.consultaHide("INSERT INTO telefonoCliente (numeroTel, idCliente) VALUES ('" & resultadosTxt & "'," & idCliente & ");")
+            ActualizarTablaTelefono()
+        End If
         ConfirmacionMensaje.soloNumBool = False
     End Sub
 
@@ -254,9 +265,10 @@ Public Class ExploradorClientes
     Private Sub btnEliminarTel_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles btnEliminarTel.Click
         Dim idTel As String
         idTel = dgvTelefono.CurrentRow.Cells(0).Value.ToString
-
-        consultas.consultaHide("DELETE FROM telefonoCliente where idTelefono=" & idTel & ";")
-        ActualizarTablaTelefono()
+        If idTel > 0 Then
+            consultas.consultaHide("DELETE FROM telefonoCliente where idTelefono=" & idTel & ";")
+            ActualizarTablaTelefono()
+        End If
     End Sub
